updated calc and calc tests to include double as instructed

+++ src/main/java/Calculator.java
@@ -13,8 +13,8 @@ public class Calculator {
         int sum = num1 * num2;
         return sum;
     }
-    public static int divide(int num1, int num2) {
-        int sum = num1 / num2;
+    public static double divide(double num1, double num2) {
+        double sum = num1 / num2;
         return sum;
     }
